pyproject.toml
[tool.poetry]
name = "osuc-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
python-environ = "^0.4.54"
python-telegram-bot = "^13.5"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
pycodestyle = "^2.7.0"
isort = "^5.10.1"
black = "^21.11b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
